+++ lib/Core/Executor.h
@@ -294,6 +294,14 @@ private:
   ref<Expr> jsontoExpr(ExecutionState &state, const MemoryObject* mo, std::string name, 
                                uint64_t relativeOffset, json &j, bool isUnderflow);
 
+ Expr::Kind getExprInfo(const ref<Expr> &e, std::string &moName, uint64_t &offset, unsigned &width );
+ bool recordWritableLocationsToJson(ExecutionState &state,const ref<Expr> &address, uint64_t address_offset);
+ void recordReadableLocationToJson(ExecutionState &state,const ref<Expr> &value);
+ bool getMoControllableInfo(ExecutionState &state, const MemoryObject* mo);
+ void detectInfomationLeak(ExecutionState &state, ref<Expr> &address, ref<Expr> &value, KInstruction *target);
+ void vulnerabilityReport(std::string report_str,KInstruction *target);
+ void detectUnintializedPointersDereferencing(ExecutionState &state, ref<Expr> &address, KInstruction *target);
+ void unitializedPointerDereferenceReport(KInstruction *target);
   /// Allocate and bind a new object in a particular state. NOTE: This
   /// function may fork.
   ///