+++ runtime/Freestanding/memset.c
@@ -11,7 +11,7 @@
 
 void *memset(void *dst, int s, size_t count) {
   char *a = dst;
-  while (count-- > 0)
+  if (count-- > 0)
     *a++ = s;
   return dst;
 }