+++ CHANGELOG.md
@@ -53,5 +53,8 @@
 - 增加--dereference-locations-input-dir=path/to/dereference_location*.json/directory的命令行选项，用于第二阶段指定输入文件目录
 
 ## [3.6.1] - 2023-9-10
-- 修改memset函数中的循环为条件判断
-- 在Json记录中增加type_name标记，方便阅读
+- 修改memset函数中的循环为条件判断。
+- 在Json记录中增加type_name标记，方便阅读。
+- 增加informationleak检查，暂未考虑controllable属性以及threadx那种情况。
+- 增加unitialized pointer dereference的检查。
+- 增加record writable和readable的信息，并将它们以memory object对象单独进行记录。